[CI][UT] move MiniMax M3 sparse attention UT to one-card e2e
